Change 'validateKitPack' behavior.
It no longer converts exceptions to 'Left's -- no point, as the caller
would probably want to catch any exceptions not covered in by type
signature, anyway. Might as well catch the lot and make it easier to
sort things out by only needing to deal with exceptions, not exceptions
+ 'Left's.

Add some tests for 'validateKitPack', as well.