Remove 'KitPath'; add tests for 'installKitPack'.
'KitPath' was a red herring. What you really want is for the user not to
have to worry about that: just specify a 'UserDirFilePath' to
'installKitPack' and it will take care of the rest.

(As a result of this, the user can't get the kit path directly. It's
likely that, in the future, we'll need some kind of 'traverseKitPacks'
functionality so that you can examine what's installed, etc.)

Also fixed a bug where 'installKitPack' failed if the "/graphics/kits"
subdirectory didn't already exist in the user directory. Now it will
create it if it doesn't already exist. However, it will *not* create the
user directory if that doesn't exist, as that indicates that either a)
the specified directory is incorrect or b) the game isn't installed.