Don't version the unpacked mod top-level directory.
Adding a file date suffix to the unpacked mod directory as a naive
versioning/manifest mechanism works fine for kit packs and other
graphics, but won't work for more complicated mods like real name fixes,
where the contents of the pack need to go into directories whose names
are prescribed by the game.

For the latter mod types, we'll need some other kind of manifest system
if we want to keep track of what's installed (and, potentially, how to
uninstall it). It will be easier just to use a single manifest system,
therefore we'll scrap this naive versioning system in favor of
the (yet-to-be-developed) more sophisticated one.