Start refactoring to support RNF mod for FM17.
My first pass tried to make the existing RealNamesFix.hs module
support both the FM16 and FM17 versions of the mod, but they're
different enough that, in the end, I think it's probably a better idea
to make separate repack modules for each. It's a bit ugly, but not as
ugly as the combined approach looked.